refactor(model): main 함수의 비즈니스 로직, 검증 로직을 별도 클래스로 분리
Application.kt에 포함되어 있던 Car 클래스, 레이싱 진행 핵심 로직, 검증 로직을
각각 model 패키지로 이동시켜 MVC 구조를 명확히 했습니다. 최종적으로 해당
프로젝트는 MVC 설계구조를 따르며, 각 계층은 아래의 기능을 수행합니다.
- controller: 흐름 제어 및 입력/출력 조합
- model: 비즈니스 로직 및 검증 로직 관리
- view: 입출력 처리
