+++ src/main/kotlin/racingcar/view/OutputView.kt
@@ -1,5 +1,5 @@
 package racingcar.view
-import racingcar.Car
+import racingcar.model.Car
 
 private const val INPUT_CAR_NAMES_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)"
 private const val INPUT_TRIAL_COUNT_MESSAGE = "시도할 횟수는 몇 회인가요?"
